Simplify heading level plugin
