+++ locales/de.js
@@ -48,9 +48,6 @@ SirTrevor.Locales.de = {
     },
     heading: {
       'title': 'Überschrift'
-    },
-    headingLevel: {
-      'title': 'Überschrift'
     }
   }
 };
