+++ locales/es.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.es = {
     },
     heading: {
       'title': 'Título'
-    },
-    headingLevel: {
-      'title': 'Título'
     }
   }
 };
