+++ locales/fi.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.fi = {
     },
     heading: {
       'title': 'Otsikko'
-    },
-    headingLevel: {
-      'title': 'Otsikko'
     }
   }
 };
