+++ locales/fr.js
@@ -48,9 +48,6 @@ SirTrevor.Locales.fr = {
     },
     heading: {
       'title': 'Titre'
-    },
-    headingLevel: {
-      'title': 'Titre'
     }
   }
 };
