+++ locales/pt-BR.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.pt_BR = {
         },
         heading: {
           'title': "Título"
-        },
-        headingLevel: {
-          'title': "Título"
         }
       }
     };
