+++ locales/pt.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.pt = {
     },
     heading: {
       'title': 'Título'
-    },
-    headingLevel: {
-      'title': 'Título'
     }
   }
 };
