+++ locales/ru.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.ru = {
     },
     heading: {
       'title': 'Заголовок'
-    },
-    headingLevel: {
-      'title': 'Заголовок'
     }
   }
 };
