+++ locales/zh-cn.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.cn = {
     },
     heading: {
       'title': '标题'
-    },
-    headingLevel: {
-      'title': '标题'
     }
   }
 }
