+++ locales/zh-tw.js
@@ -46,9 +46,6 @@ SirTrevor.Locales.cn_TW = {
     },
     heading: {
       'title': '標題'
-    },
-    headingLevel: {
-      'title': '標題'
     }
   }
 }
