+++ src/blocks/heading.js
@@ -9,7 +9,7 @@ var stToHTML = require('../to-html');
 
 var ScribeTextBlockPlugin = require('./scribe-plugins/scribe-text-block-plugin');
 var ScribeQuotePlugin = require('./scribe-plugins/scribe-quote-plugin');
-var ScribeHeadingLevelPlugin = require('./scribe-plugins/scribe-heading-level-plugin');
+var ScribeHeadingPlugin = require('./scribe-plugins/scribe-heading-plugin');
 
 module.exports = Block.extend({
 
@@ -18,7 +18,7 @@ module.exports = Block.extend({
   editorHTML: '<h2 class="st-required st-text-block st-text-block--heading" contenteditable="true"></h2>',
 
   configureScribe: function(scribe) {
-    scribe.use(new ScribeHeadingLevelPlugin(this));
+    scribe.use(new ScribeHeadingPlugin(this));
     scribe.use(new ScribeTextBlockPlugin(this));
     scribe.use(new ScribeQuotePlugin(this));
 
@@ -39,14 +39,13 @@ module.exports = Block.extend({
   icon_name: 'heading',
 
   loadData: function(data) {
-    console.log('LOAD DATA', data)
     if (this.options.convertFromMarkdown && data.format !== "html") {
       this.setTextBlockHTML(stToHTML(data.text, this.type));
     } else {
       this.setTextBlockHTML(data.text);
     }
 
-    data.level = data.level || 2;
+    data.level = data.level || this.editorOptions.defaultHeadingLevel;
     this.el.dataset.level = data.level;
   },
