+++ src/blocks/scribe-plugins/scribe-heading-plugin.js
@@ -2,23 +2,38 @@
 
 var scribeHeadingPlugin = function(block) {
   return function(scribe) {
-    
-    const headingCommand = new scribe.api.Command('heading');
+
+    const { defaultHeadingLevel, headingLevels } = block.editorOptions;
+    const minHeadingLevel = headingLevels[0];
+    const maxHeadingLevel = headingLevels[headingLevels.length - 1];
+
+    const headingCommand = new scribe.api.Command(`heading`);
     headingCommand.queryEnabled = () => {
       return block.inline_editable;
     };
     headingCommand.queryState = () => {
-      return block.type === 'heading';
-    };
-
-    const getBlockType = function() {
-      return headingCommand.queryState() ? 'Text' : 'Heading';
+      if (block.type === 'heading') {
+        return block.getBlockData().level || defaultHeadingLevel || minHeadingLevel;
+      } else {
+        return false;
+      }
     };
 
     headingCommand.execute = function headingCommandExecute(value) {
-      var blockType = getBlockType()
+      var level = block.getBlockData().level + 1;
+      var blockType = 'Heading';
+
+      if (!level || level < minHeadingLevel) {
+        level = minHeadingLevel;
+      } else if (level > maxHeadingLevel) {
+        level = null;
+        blockType = 'Text';
+      }
+
+
       var data = {
         format: 'html',
+        level: level,
         text: block.getScribeInnerContent()
       };
